+++ tax/src/main/java/pt/ulisboa/tecnico/softeng/tax/services/remote/dataobjects/RestInvoiceData.java
@@ -13,17 +13,17 @@ public class RestInvoiceData {
 	private String sellerNif;
 	private String buyerNif;
 	private String itemType;
-	private Double value;
+	private Long value;
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
 	private LocalDate date;
-	private Double iva;
+	private Long iva;
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss.SSS")
 	private DateTime time;
 
 	public RestInvoiceData() {
 	}
 
-	public RestInvoiceData(String reference, String sellerNif, String buyerNif, String itemType, Double value,
+	public RestInvoiceData(String reference, String sellerNif, String buyerNif, String itemType, long value,
 			LocalDate date, DateTime time) {
 		if (reference == null) {
 			throw new TaxException();
@@ -32,7 +32,7 @@ public class RestInvoiceData {
 		this.sellerNif = sellerNif;
 		this.buyerNif = buyerNif;
 		this.itemType = itemType;
-		this.value = value;
+		this.value =  value;
 		this.date = date;
 		this.time = time;
 	}
@@ -80,14 +80,14 @@ public class RestInvoiceData {
 		this.itemType = itemType;
 	}
 
-	public Double getValue() {
+	public Long getValue() {
 		return this.value;
 	}
 
-	public void setValue(Double value) {
+	public void setValue(Long value) {
 		this.value = value;
 	}
-
+	
 	public LocalDate getDate() {
 		return this.date;
 	}
@@ -96,11 +96,11 @@ public class RestInvoiceData {
 		this.date = date;
 	}
 
-	public Double getIva() {
+	public Long getIva() {
 		return this.iva;
 	}
 
-	public void setIva(Double iva) {
+	public void setIva(Long iva) {
 		this.iva = iva;
 	}
